chore: move multisig into wallet namespace
