+++ ironfish/src/rpc/routes/index.ts
@@ -10,7 +10,6 @@ export * from './peer'
 export * from './router'
 export * from './rpc'
 export * from './miner'
-export * from './multisig'
 export * from './faucet'
 export * from './worker'
 export * from './mempool'
