+++ ironfish/src/rpc/routes/namespaces.ts
@@ -8,7 +8,6 @@ export enum ApiNamespace {
   event = 'event',
   faucet = 'faucet',
   miner = 'miner',
-  multisig = 'multisig',
   node = 'node',
   peer = 'peer',
   wallet = 'wallet',
