+++ ironfish/src/rpc/routes/wallet/index.ts
@@ -26,6 +26,7 @@ export * from './getNotes'
 export * from './getPublicKey'
 export * from './importAccount'
 export * from './mintAsset'
+export * from './multisig'
 export * from './postTransaction'
 export * from './remove'
 export * from './removeAccount'
