+++ ironfish/src/rpc/routes/wallet/multisig/createSigningCommitment.test.ts
@@ -2,16 +2,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { ParticipantSecret } from '@ironfish/rust-nodejs'
-import { createRouteTest } from '../../../testUtilities/routeTest'
+import { createRouteTest } from '../../../../testUtilities/routeTest'
 
-describe('Route multisig/createSigningCommitment', () => {
+describe('Route wallet/multisig/createSigningCommitment', () => {
   const routeTest = createRouteTest()
 
   it('should error on invalid keypackage', async () => {
     const keyPackage = 'invalid key package'
     const request = { keyPackage, seed: 0 }
     await expect(
-      routeTest.client.request('multisig/createSigningCommitment', request).waitForEnd(),
+      routeTest.client.request('wallet/multisig/createSigningCommitment', request).waitForEnd(),
     ).rejects.toThrow('InvalidData')
   })
 
@@ -21,12 +21,14 @@ describe('Route multisig/createSigningCommitment', () => {
     }))
 
     const request = { minSigners: 2, participants }
-    const response = await routeTest.client.multisig.createTrustedDealerKeyPackage(request)
+    const response = await routeTest.client.wallet.multisig.createTrustedDealerKeyPackage(
+      request,
+    )
 
     const trustedDealerPackage = response.content
 
     const signingCommitment = await routeTest.client
-      .request('multisig/createSigningCommitment', {
+      .request('wallet/multisig/createSigningCommitment', {
         keyPackage: trustedDealerPackage.keyPackages[0].keyPackage,
         seed: 420,
       })
