+++ ironfish/src/rpc/routes/wallet/multisig/createSigningPackage.ts
@@ -3,8 +3,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { UnsignedTransaction } from '@ironfish/rust-nodejs'
 import * as yup from 'yup'
-import { ApiNamespace } from '../namespaces'
-import { routes } from '../router'
+import { ApiNamespace } from '../../namespaces'
+import { routes } from '../../router'
 import { RpcSigningCommitments, RpcSigningCommitmentsSchema } from './types'
 
 export type CreateSigningPackageRequest = {
@@ -32,7 +32,7 @@ export const CreateSigningPackageResponseSchema: yup.ObjectSchema<CreateSigningP
     .defined()
 
 routes.register<typeof CreateSigningPackageRequestSchema, CreateSigningPackageResponse>(
-  `${ApiNamespace.multisig}/createSigningPackage`,
+  `${ApiNamespace.wallet}/multisig/createSigningPackage`,
   CreateSigningPackageRequestSchema,
   (request, _context): void => {
     const unsignedTransaction = new UnsignedTransaction(
