+++ ironfish/src/rpc/routes/wallet/multisig/createSigningShare.ts
@@ -3,8 +3,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { createSigningShare, UnsignedTransaction } from '@ironfish/rust-nodejs'
 import * as yup from 'yup'
-import { ApiNamespace } from '../namespaces'
-import { routes } from '../router'
+import { ApiNamespace } from '../../namespaces'
+import { routes } from '../../router'
 
 export type CreateSigningShareRequest = {
   signingPackage: string
@@ -34,7 +34,7 @@ export const CreateSigningShareResponseSchema: yup.ObjectSchema<CreateSigningSha
     .defined()
 
 routes.register<typeof CreateSigningShareRequestSchema, CreateSigningShareResponse>(
-  `${ApiNamespace.multisig}/createSigningShare`,
+  `${ApiNamespace.wallet}/multisig/createSigningShare`,
   CreateSigningShareRequestSchema,
   (request, _context): void => {
     const unsigned = new UnsignedTransaction(
